Move test classes to its own file
