+++ src/libraries/System.Text.Json/System.Text.Json.SourceGeneration.Tests/System.Text.Json.SourceGeneration.Tests.csproj
@@ -8,6 +8,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Compile Include="TestClasses.cs" />
     <Compile Include="JsonSourceGeneratorTests.cs" />
   </ItemGroup>
